backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-polymorphic==3.0.0
djangorestframework==3.12.2
drf-nested-routers==0.92.5
gunicorn==20.0.4
Markdown==3.3.3
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
django-polymorphic==3.0.0
six==1.15.0
requests==2.25.1
httpx==0.16.1
